requirements.txt
langchain-community
langchain-google-genai
langchain-core
langchain-text-splitters
langchain-huggingface
python-dotenv
sentence-transformers
faiss-cpu
